validacoes, readme e identação

+++ src/controllers/atendimentosController.js
@@ -27,13 +27,19 @@ class atendimentosController {
     
     try {
 
-      const umRegistro = await database.findOne({ 
-        where: {
-          id: id
-        }
-      })
+      const isValid = Validacoes.validaId(id)
+
+      if(await isValid) {
+
+      const umRegistro = await database.findOne({where: { id: id}})
 
       return res.status(200).json(umRegistro)
+
+    } else {
+
+      throw new Error ("Insira um número e um id de atendimento existente.")
+
+    }
 
     } catch (error) {
 
@@ -67,6 +73,7 @@ class atendimentosController {
     const regModificado = req.body
 
     try {
+      
       const isValid = Validacoes.validaId(id)
 
       if(await isValid) {
@@ -80,9 +87,8 @@ class atendimentosController {
       } else {
 
         throw new Error ("Insira um número e um id de atendimento existente.")
-      }
 
-    
+      }
 
     } catch (error) {
 
@@ -98,13 +104,23 @@ class atendimentosController {
 
     try {
 
-      await database.destroy({where: {id:id}})
+      const isValid = Validacoes.validaId(id)
 
-      return res.status(200).json({mensagem: `O id ${id} foi deletado com sucesso.`})
+      if(await isValid) {
+
+        await database.destroy({where: {id:id}})
+
+        return res.status(200).json({mensagem: `O id ${id} foi deletado com sucesso.`})
+
+      } else {
+
+        throw new Error ("Insira um número e um id de atendimento existente.")
+
+      }
 
     } catch (error) {
 
-      return res.status(400).json(error.message)
+        return res.status(400).json(error.message)
 
     }
